+++ tools/swing/robot/src/main/java/is/codion/tools/swing/robot/DefaultAutomation.java
@@ -75,12 +75,8 @@ final class DefaultAutomation implements Automation {
 		if (builder.applicationWindow == null) {
 			return null;
 		}
-		Narrator narrator = Narrator.narrator(controller);
-		if (builder.applicationWindow != null) {
-			narrator.attach(builder.applicationWindow);
-		}
 
-		return narrator;
+		return Narrator.narrator(controller, builder.applicationWindow);
 	}
 
 	private static void handleException(Exception exception) {